update 210531 18:00

+++ src/main/java/com/candlebe/gcoach/controller/SettingController.java
@@ -14,6 +14,8 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 
+import java.util.Optional;
+
 @Controller
 @Log4j2
 @RequiredArgsConstructor
@@ -40,15 +42,21 @@ public class SettingController {
     }
 
     @PostMapping("/deleteAccount")
-    public String postDeleteAccount(@AuthenticationPrincipal AuthMemberDTO authMemberDTO) {
-        Member member = memberRepository.findByUsername(authMemberDTO.getUsername(), authMemberDTO.isFormSocial()).orElseThrow();
-        historyRepository.deleteHistories(member);
-        likeRepository.deleteLikes(member);
-        replyRepository.deleteReplies(member);
-        diaryRepository.deleteDiaries(member);
-        member.getRoleSet().clear();
-        memberRepository.save(member);
-        memberRepository.deleteMember(member.getMid());
+    public String postDeleteAccount(@AuthenticationPrincipal AuthMemberDTO authMemberDTO, Model model) {
+        Optional<Member> members = memberRepository.findByUsername(authMemberDTO.getUsername(), authMemberDTO.isFormSocial());
+        if (members.isPresent()) {
+            Member member = members.get();
+            historyRepository.deleteHistories(member);
+            likeRepository.deleteLikes(member);
+            replyRepository.deleteReplies(member);
+            diaryRepository.deleteDiaries(member);
+            member.getRoleSet().clear();
+            memberRepository.save(member);
+            memberRepository.deleteMember(member.getMid());
+            model.addAttribute("msg", "회원탈퇴 성공");
+            return "redirect:/logout";
+        }
+        model.addAttribute("msg", "회원탈퇴 실패");
         return "redirect:/logout";
     }
 }
